+++ client-ts/src/tests/App.test.tsx
@@ -1,8 +1,7 @@
 import React, { ReactNode } from 'react';
 import { render, screen } from '@testing-library/react';
-import { MemoryRouter } from 'react-router-dom';
-import App from './App';
-import Login from './pages/Login';
+import App from '../App';
+import Login from '../pages/Login';
 
 jest.mock('swiper/react', () => ({
   Swiper: ({ children }: { children?: ReactNode }) => <div>{children}</div>,
@@ -21,16 +20,3 @@ test('renders app', () => {
   expect(linkElement).toBeInTheDocument();
 });
 
-test('render login', () => {
-  render(
-    <MemoryRouter>
-      <Login />
-    </MemoryRouter>
-  );
-
-  const getLogin = screen.getByText(/Login/i);
-  const getRegister = screen.getByText(/Register/i);
-
-  expect(getLogin).toBeInTheDocument();
-  expect(getRegister).toBeInTheDocument();
-});
